edit entry modal created for modularity
