+++ client/src/utils/quillConfig.js
@@ -1,0 +1,32 @@
+// Quill modules and formats configuration
+export const quillModules = {
+    toolbar: [
+        [{ 'header': [1, 2, 3, false] }],
+        ['bold', 'italic', 'underline'],
+        [{ 'list': 'ordered' }, { 'list': 'bullet' }],
+        [{ 'align': ['', 'center', 'right', 'justify'] }],
+        ['clean']
+    ],
+    clipboard: {
+        matchVisual: false
+    },
+    keyboard: {
+        bindings: {
+            tab: {
+                key: 9,
+                handler: function () {
+                    return true; // Let default tab behavior happen
+                }
+            }
+        }
+    }
+};
+
+export const quillFormats = [
+    'header',
+    'bold', 'italic', 'underline',
+    'list', 'bullet', 'ordered',
+    'align',
+    'indent',
+    'direction'
+];
